REP-114 refactored ID generator and resolved dependencies
