+++ .gitignore
@@ -16,3 +16,4 @@
 
 /.idea/
 *.iml
+**/out
