+++ build.gradle
@@ -37,7 +37,9 @@ subprojects  {
         jsonassertVersion = '1.5.0'
         apiCoreVersion = '1.9.4_873-RC'
         commonsCollectionsVersion = '4.1'
+        hibernateValidator = '6.0.15.Final'
         javaxValidationVersion = '2.0.1.Final'
+        javaxELVersion = '3.0.1-b09'
         springFoxSwaggerVersion = '2.9.2'
         db2DriverVersion = '4.0.0'
         junitVersion = '4.12'
