+++ cares-common/build.gradle
@@ -8,9 +8,21 @@ apply plugin: 'eclipse'
 sourceCompatibility = 1.8
 
 dependencies {
+    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:' + jacksonVersion
+    implementation 'gov.ca.cwds.api:api-core-common:' + apiCoreVersion, {
+        transitive = false
+    }
+    implementation 'javax.validation:validation-api:' + javaxValidationVersion
+    implementation 'org.apache.commons:commons-collections4:' + commonsCollectionsVersion
+    implementation 'org.apache.commons:commons-lang3:' + commonsLangVersion
+    implementation 'org.hibernate:hibernate-validator:' + hibernateValidator
+    implementation 'org.glassfish:javax.el:' + javaxELVersion
 
+    implementation 'org.springframework.boot:spring-boot-starter-aop:' + springBootVersion
+    implementation 'org.apache.commons:commons-lang3:' + commonsLangVersion
     implementation 'org.springframework.boot:spring-boot:' + springBootVersion
-    implementation 'org.springframework.boot:spring-boot-starter-aop:' + springBootVersion   
-    
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:' + jacksonVersion       
+
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.hamcrest:hamcrest:2.1'
+    testImplementation "org.mockito:mockito-core:2.25.0"
 }
