+++ cares-common/src/main/java/gov/ca/cwds/cares/common/identifier/Base62.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.data.persistence.cms;
+package gov.ca.cwds.cares.common.identifier;
 
 /**
  * Common base-62 algorithm. Varied flavors online. Customized for CMS legacy keys. Converts base-10
