+++ cares-common/src/main/java/gov/ca/cwds/cares/common/identifier/CmsKeyIdGenerator.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.data.persistence.cms;
+package gov.ca.cwds.cares.common.identifier;
 
 import java.io.Serializable;
 import java.util.Calendar;
