+++ cares-common/src/test/java/gov/ca/cwds/cares/common/identifier/Base62Test.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.data.persistence.cms;
+package gov.ca.cwds.cares.common.identifier;
 
 import static junit.framework.Assert.assertEquals;
